manually add serde bindings for fbs models
