+++ bd-proto/build.rs
@@ -115,7 +115,7 @@ fn main() {
     ],
     includes: &[Path::new("../api/src")],
     out_dir: Path::new("src/flatbuffers"),
-    extra: &["--gen-object-api", "--rust-serialize"],
+    extra: &["--gen-object-api"],
     ..flatc_rust::Args::default()
   })
   .unwrap();
@@ -126,7 +126,7 @@ fn main() {
     ],
     includes: &[Path::new("../api/src")],
     out_dir: Path::new("src/flatbuffers"),
-    extra: &["--gen-object-api", "--rust-serialize"],
+    extra: &["--gen-object-api"],
     ..flatc_rust::Args::default()
   })
   .unwrap();
